Remove InterpolationType support, there is no need for it.
