+++ interpolator/cpp/include/interpolator/BaseInterpolator.hpp
@@ -20,7 +20,7 @@ class BaseInterpolator : public IInterpolator
 
     virtual ~BaseInterpolator() = default;
 
-    BaseInterpolator(const Vertices &trajectory, const InterpolationType &interpolation_type);
+    BaseInterpolator(const Vertices &trajectory);
 
     const Vertices &trajectory() const final;
     void set_trajectory(const Vertices &trajectory) final;
@@ -36,8 +36,6 @@ class BaseInterpolator : public IInterpolator
     double x_at_position(double position) const final;
     double y_at_position(double position) const final;
     double z_at_position(double position) const final;
-    InterpolationType interpolation_type() const final;
-    std::string interpolation_type_str() const final;
 
   private:
     typedef double (BaseInterpolator::*DeltaCalculator)(double, const AdjacentVertices &) const;
@@ -186,9 +184,6 @@ class BaseInterpolator : public IInterpolator
 
   private:
     Vertices _trajectory;
-
-  protected:
-    const InterpolationType _interpolation_type;
 };
 
 } // namespace splines
