+++ interpolator/cpp/include/interpolator/CubicInterpolator.hpp
@@ -14,7 +14,7 @@ class CubicInterpolator : public BaseInterpolator
 {
   public:
     CubicInterpolator(const Vertices &trajectory)
-        : BaseInterpolator(trajectory, InterpolationType::cubic)
+        : BaseInterpolator(trajectory)
     {
     }
